OPE-11 split core tools by domain
