+++ src/opentulpa/agent/tools/core_tools.py
@@ -2,39 +2,16 @@
 
 from __future__ import annotations
 
-import asyncio
 import json
 import re
 from contextlib import suppress
-from datetime import UTC, datetime
 from typing import Any
-from urllib.parse import urlparse
 
-import httpx
-from langchain.tools import tool
-
-from opentulpa.agent.file_analysis import summarize_uploaded_blob
-from opentulpa.agent.lc_messages import HumanMessage, SystemMessage
 from opentulpa.agent.tools.common import (
-    normalize_command_for_working_dir,
-    normalize_execution_origin,
     require_customer_id,
     require_thread_id,
 )
-from opentulpa.agent.utils import content_to_text as _content_to_text
-from opentulpa.agent.utils import extract_html_title as _extract_html_title
 from opentulpa.agent.utils import html_to_text as _html_to_text
-from opentulpa.agent.utils import looks_like_shell_command as _looks_like_shell_command
-from opentulpa.context.customer_profile_models import (
-    CustomerScopedClearResponse,
-    CustomerScopedOkResponse,
-    CustomerScopedRequest,
-    DirectiveGetResponse,
-    DirectiveSetRequest,
-    TimeProfileGetResponse,
-    TimeProfileSetRequest,
-    TimeProfileSetResponse,
-)
 
 
 def _tool_error_payload(tool_name: str, response: Any) -> dict[str, Any]:
@@ -493,962 +470,29 @@ async def _sync_proactive_heartbeat(
         "schedule": f"0 */{interval_hours} * * *",
     }
 
-
 def register_core_tools(runtime: Any) -> dict[str, Any]:
-    @tool
-    async def send_owner_update(message: str, dedupe_key: str = "") -> Any:
-        """Send a short interim update to the current owner/support Telegram chat.
+    from opentulpa.agent.tools.business_knowledge_tools import register_business_knowledge_tools
+    from opentulpa.agent.tools.directive_tools import register_directive_tools
+    from opentulpa.agent.tools.file_tools import register_file_tools
+    from opentulpa.agent.tools.memory_tools import register_memory_tools
+    from opentulpa.agent.tools.owner_update_tools import register_owner_update_tools
+    from opentulpa.agent.tools.server_time_tools import register_server_time_tools
+    from opentulpa.agent.tools.task_tools import register_task_tools
+    from opentulpa.agent.tools.time_profile_tools import register_time_profile_tools
+    from opentulpa.agent.tools.tulpa_workspace_tools import register_tulpa_workspace_tools
+    from opentulpa.agent.tools.user_context_tools import register_user_context_tools
+    from opentulpa.agent.tools.web_tools import register_web_tools
 
-        Use only during live owner/support turns when you will continue working
-        with tools. This is for long-running interactive or workflow setup work,
-        not final answers, inbound lead replies, routine wakes, or background
-        event notifications.
-        """
-        require_customer_id(runtime)
-        require_thread_id(runtime)
-        safe_message = str(message or "").strip()
-        if not safe_message:
-            return {"ok": False, "sent": False, "reason": "empty_message"}
-        if len(safe_message) > 500:
-            safe_message = safe_message[:497].rstrip() + "..."
-        emitter = getattr(runtime, "emit_interactive_update", None)
-        if not callable(emitter):
-            return {"ok": False, "sent": False, "reason": "interactive_update_unavailable"}
-        return await emitter(
-            text=safe_message,
-            dedupe_key=str(dedupe_key or "").strip(),
-        )
-
-    @tool
-    async def memory_search(query: str) -> Any:
-        """Search user memory."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/memory/search",
-            json_body={"query": query, "user_id": customer_id, "limit": 5},
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"memory_search failed: {r.text}"}
-        return r.json().get("results", [])
-
-    @tool
-    async def memory_add(summary: str) -> Any:
-        """Store a user memory summary."""
-        customer_id = require_customer_id(runtime)
-        retryable_errors = (
-            httpx.ConnectError,
-            httpx.ConnectTimeout,
-            httpx.ReadTimeout,
-            httpx.RemoteProtocolError,
-            RuntimeError,
-        )
-        for attempt in range(2):
-            try:
-                r = await runtime._request_with_backoff(
-                    "POST",
-                    "/internal/memory/add",
-                    json_body={
-                        "messages": [{"role": "user", "content": summary}],
-                        "user_id": customer_id,
-                    },
-                    timeout=30.0,
-                    retries=0,
-                )
-                if r.status_code != 200:
-                    return {"error": f"memory_add failed ({r.status_code}): {r.text}"}
-                return {"ok": True}
-            except retryable_errors as exc:
-                if attempt == 0:
-                    await asyncio.sleep(1.5)
-                    continue
-                exc_name = type(exc).__name__
-                detail = str(exc) or exc_name
-                return {"error": f"memory_add timed out after retries: {detail}"}
-        return {"error": "memory_add failed: exhausted retries"}
-
-    @tool
-    async def uploaded_file_search(query: str, limit: int = 5) -> Any:
-        """Search uploaded files for this user by natural-language query."""
-        customer_id = require_customer_id(runtime)
-        safe_limit = max(1, min(int(limit), 20))
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/search",
-            json_body={
-                "query": query,
-                "customer_id": customer_id,
-                "limit": safe_limit,
-            },
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"uploaded_file_search failed: {r.text}"}
-        return r.json().get("results", [])
-
-    @tool
-    async def uploaded_file_get(
-        file_id: str,
-        max_excerpt_chars: int = 16000,
-    ) -> Any:
-        """Get metadata and text excerpt for one uploaded file."""
-        customer_id = require_customer_id(runtime)
-        safe_chars = max(500, min(int(max_excerpt_chars), 60000))
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/get",
-            json_body={
-                "file_id": file_id,
-                "customer_id": customer_id,
-                "max_excerpt_chars": safe_chars,
-            },
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"uploaded_file_get failed: {r.text}"}
-        return r.json().get("file", {})
-
-    @tool
-    async def uploaded_file_send(
-        file_id: str,
-        caption: str | None = None,
-    ) -> Any:
-        """Send a previously uploaded file back to the user's Telegram chat."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/send",
-            json_body={
-                "file_id": file_id,
-                "customer_id": customer_id,
-                "caption": caption,
-            },
-            timeout=25.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"uploaded_file_send failed: {r.text}"}
-        return _with_delivery_instruction(r.json())
-
-    @tool
-    async def tulpa_file_send(
-        path: str,
-        caption: str | None = None,
-    ) -> Any:
-        """Send a local file from tulpa_stuff/ back to the user's Telegram chat."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/send_local",
-            json_body={
-                "path": path,
-                "customer_id": customer_id,
-                "caption": caption,
-            },
-            timeout=25.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"tulpa_file_send failed: {r.text}"}
-        return _with_delivery_instruction(r.json())
-
-    @tool
-    async def web_image_send(
-        url: str,
-        caption: str | None = None,
-        max_bytes: int = 10_000_000,
-    ) -> Any:
-        """Download an image from a web URL and send it to Telegram."""
-        customer_id = require_customer_id(runtime)
-        safe_max_bytes = max(250_000, min(int(max_bytes), 25_000_000))
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/send_web_image",
-            json_body={
-                "url": url,
-                "customer_id": customer_id,
-                "caption": caption,
-                "max_bytes": safe_max_bytes,
-            },
-            timeout=70.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return {"error": f"web_image_send failed: {r.text}"}
-        return _with_delivery_instruction(r.json())
-
-    @tool
-    async def uploaded_file_analyze(
-        file_id: str,
-        question: str | None = None,
-    ) -> Any:
-        """Analyze a previously uploaded file again, optionally with a focused question.
-
-        Do not use this as a fallback for workflow knowledge files that should be
-        indexed and queried with business_knowledge_index/business_knowledge_query.
-        """
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/analyze",
-            json_body={
-                "file_id": file_id,
-                "customer_id": customer_id,
-                "question": question,
-            },
-            timeout=60.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return {"error": f"uploaded_file_analyze failed: {r.text}"}
-        return r.json()
-
-    @tool
-    async def uploaded_file_inspect_structure(
-        file_id: str,
-        search_terms: list[str] | str | None = None,
-    ) -> Any:
-        """Inspect an uploaded file's structure before selecting workflow knowledge.
-
-        Use this first for arbitrary spreadsheets or large source files. For XLSX files,
-        it opens the workbook, returns sheet names, dimensions, sample rows, table
-        candidates, and optional matches for search_terms derived from the user's
-        workflow goal. For workflow knowledge, prefer business_knowledge_index and
-        business_knowledge_query so the source file stays out of chat context.
-        """
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/files/inspect_structure",
-            json_body={
-                "file_id": file_id,
-                "customer_id": customer_id,
-                "search_terms": search_terms,
-            },
-            timeout=60.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("uploaded_file_inspect_structure", r)
-        return _compact_uploaded_file_inspection(r.json())
-
-    @tool
-    async def business_knowledge_index(
-        file_ids: list[str],
-        scope_type: str = "current_workflow",
-        scope_id: str = "",
-    ) -> Any:
-        """Prepare uploaded source files as scoped business knowledge.
-
-        Use this during workflow setup when uploaded files should become durable
-        source knowledge for the workflow. If no setup session exists yet, call
-        intake_workflow_setup_begin first. This normalizes source files into an
-        LLM-readable knowledge pack. Bind original uploaded source file ids to
-        draft_patch.knowledge_file_ids; do not create a summarized Markdown pack.
-        """
-        customer_id = require_customer_id(runtime)
-        safe_file_ids = [
-            str(item or "").strip()
-            for item in list(file_ids or [])
-            if str(item or "").strip()
-        ][:20]
-        if not safe_file_ids:
-            return {"error": "business_knowledge_index failed: file_ids is required"}
-        resolved = await _resolve_business_knowledge_scope(
-            runtime,
-            scope_type=scope_type,
-            scope_id=scope_id,
-        )
-        if isinstance(resolved, dict):
-            return resolved
-        resolved_type, resolved_id = resolved
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/knowledge/index_sources",
-            json_body={
-                "customer_id": customer_id,
-                "scope_type": resolved_type,
-                "scope_id": resolved_id,
-                "file_ids": safe_file_ids,
-            },
-            timeout=60.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("business_knowledge_index", r)
-        return _compact_business_knowledge_index(r.json())
-
-    @tool
-    async def business_knowledge_query(
-        query: str,
-        scope_type: str = "current_workflow",
-        scope_id: str = "",
-    ) -> Any:
-        """Ask scoped business knowledge for source-backed facts.
-
-        Use this for source-backed business details during workflow setup or intake.
-        The oracle reads the full prepared knowledge pack for the resolved scope and
-        returns only a compact plain-text answer. During intake, do not call it again
-        just to revalidate source-backed facts already present on the active booking.
-        """
-        customer_id = require_customer_id(runtime)
-        safe_query = str(query or "").strip()
-        if not safe_query:
-            return {"error": "business_knowledge_query failed: query is required"}
-        resolved = await _resolve_business_knowledge_scope(
-            runtime,
-            scope_type=scope_type,
-            scope_id=scope_id,
-        )
-        if isinstance(resolved, dict):
-            return resolved
-        resolved_type, resolved_id = resolved
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/knowledge/query",
-            json_body={
-                "customer_id": customer_id,
-                "scope_type": resolved_type,
-                "scope_id": resolved_id,
-                "query": safe_query,
-                "max_extract_chars": 3000,
-            },
-            timeout=60.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("business_knowledge_query", r)
-        return _compact_business_knowledge_query(r.json())
-
-    @tool
-    async def user_context_add_files(file_ids: list[str]) -> Any:
-        """Add uploaded files to the durable interactive user context.
-
-        Use this only when the user's recent instructions clearly ask to add files
-        to their reusable context. If intent is unclear, ask what to do with the
-        files instead of guessing from filenames or content.
-        """
-        customer_id = require_customer_id(runtime)
-        safe_file_ids = [
-            str(item or "").strip()
-            for item in list(file_ids or [])
-            if str(item or "").strip()
-        ][:50]
-        if not safe_file_ids:
-            return {"error": "user_context_add_files failed: file_ids is required"}
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/add_files",
-            json_body={"customer_id": customer_id, "file_ids": safe_file_ids},
-            timeout=90.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_add_files", r)
-        return r.json()
-
-    @tool
-    async def user_context_query(query: str, max_extract_chars: int = 3000) -> Any:
-        """Query the durable interactive user context for grounded evidence."""
-        customer_id = require_customer_id(runtime)
-        safe_query = str(query or "").strip()
-        if not safe_query:
-            return {"error": "user_context_query failed: query is required"}
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/query",
-            json_body={
-                "customer_id": customer_id,
-                "query": safe_query,
-                "max_extract_chars": max(500, min(int(max_extract_chars), 5000)),
-            },
-            timeout=70.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_query", r)
-        return r.json()
-
-    @tool
-    async def user_context_list_sources(include_archived: bool = False) -> Any:
-        """List files that are currently registered in the interactive user context."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/list_sources",
-            json_body={"customer_id": customer_id, "include_archived": bool(include_archived)},
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_list_sources", r)
-        return r.json().get("sources", [])
-
-    @tool
-    async def user_context_find_sources(query: str, limit: int = 10) -> Any:
-        """Find user-context sources by filename, summary, or extracted text preview."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/find_sources",
-            json_body={
-                "customer_id": customer_id,
-                "query": query,
-                "limit": max(1, min(int(limit), 50)),
-            },
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_find_sources", r)
-        return r.json().get("sources", [])
-
-    @tool
-    async def user_context_reindex(file_ids: list[str] | None = None) -> Any:
-        """Reindex selected user-context files, or every active source when omitted."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/reindex",
-            json_body={"customer_id": customer_id, "file_ids": file_ids or None},
-            timeout=90.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_reindex", r)
-        return r.json()
-
-    @tool
-    async def user_context_archive_sources(file_ids: list[str]) -> Any:
-        """Archive selected files so default user-context queries stop using them."""
-        customer_id = require_customer_id(runtime)
-        safe_file_ids = [
-            str(item or "").strip()
-            for item in list(file_ids or [])
-            if str(item or "").strip()
-        ][:50]
-        if not safe_file_ids:
-            return {"error": "user_context_archive_sources failed: file_ids is required"}
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/archive_sources",
-            json_body={"customer_id": customer_id, "file_ids": safe_file_ids},
-            timeout=20.0,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_archive_sources", r)
-        return r.json()
-
-    @tool
-    async def user_context_promote_to_intake(workflow_id: str, file_ids: list[str]) -> Any:
-        """Index selected user-context files into an existing intake workflow scope."""
-        customer_id = require_customer_id(runtime)
-        safe_file_ids = [
-            str(item or "").strip()
-            for item in list(file_ids or [])
-            if str(item or "").strip()
-        ][:50]
-        if not str(workflow_id or "").strip():
-            return {"error": "user_context_promote_to_intake failed: workflow_id is required"}
-        if not safe_file_ids:
-            return {"error": "user_context_promote_to_intake failed: file_ids is required"}
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/user_context/promote_to_intake",
-            json_body={
-                "customer_id": customer_id,
-                "workflow_id": str(workflow_id).strip(),
-                "file_ids": safe_file_ids,
-            },
-            timeout=90.0,
-            retries=1,
-        )
-        if r.status_code != 200:
-            return _tool_error_payload("user_context_promote_to_intake", r)
-        return r.json()
-
-    @tool
-    async def directive_get() -> Any:
-        """Get the active persistent directive profile for this user."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/directive/get",
-            json_body=CustomerScopedRequest(customer_id=customer_id).model_dump(mode="json"),
-            timeout=5.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"directive_get failed: {r.text}"}
-        return DirectiveGetResponse.model_validate(r.json()).model_dump(mode="json")
-
-    @tool
-    async def directive_set(directive: str) -> Any:
-        """Set or overwrite the user's persistent directive profile."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/directive/set",
-            json_body=DirectiveSetRequest(
-                customer_id=customer_id,
-                directive=directive,
-                source="langgraph_tool",
-            ).model_dump(mode="json"),
-            timeout=5.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"directive_set failed: {r.text}"}
-        payload = CustomerScopedOkResponse.model_validate(r.json()).model_dump(mode="json")
-        payload["proactive_heartbeat"] = await _sync_proactive_heartbeat(
-            runtime=runtime,
-            customer_id=customer_id,
-            directive_text=directive,
-        )
-        return payload
-
-    @tool
-    async def directive_clear() -> Any:
-        """Clear the user's persistent directive profile."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/directive/clear",
-            json_body=CustomerScopedRequest(customer_id=customer_id).model_dump(mode="json"),
-            timeout=5.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"directive_clear failed: {r.text}"}
-        payload = CustomerScopedClearResponse.model_validate(r.json()).model_dump(mode="json")
-        payload["proactive_heartbeat"] = await _sync_proactive_heartbeat(
-            runtime=runtime,
-            customer_id=customer_id,
-            directive_text="disable proactive mode",
-        )
-        return payload
-
-    @tool
-    async def time_profile_get() -> Any:
-        """Get stored user UTC offset (if known)."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/time_profile/get",
-            json_body=CustomerScopedRequest(customer_id=customer_id).model_dump(mode="json"),
-            timeout=5.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"time_profile_get failed: {r.text}"}
-        return TimeProfileGetResponse.model_validate(r.json()).model_dump(mode="json")
-
-    @tool
-    async def time_profile_set(utc_offset: str) -> Any:
-        """Set user UTC offset in +HH:MM or -HH:MM format."""
-        customer_id = require_customer_id(runtime)
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/time_profile/set",
-            json_body=TimeProfileSetRequest(
-                customer_id=customer_id,
-                utc_offset=utc_offset,
-                source="langgraph_tool",
-            ).model_dump(mode="json"),
-            timeout=5.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"time_profile_set failed: {r.text}"}
-        return TimeProfileSetResponse.model_validate(r.json()).model_dump(mode="json")
-
-    @tool
-    async def web_search(query: str) -> Any:
-        """Search the web for current information."""
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/web_search",
-            json_body={"query": query},
-            timeout=90.0,
-        )
-        if r.status_code != 200:
-            return {"error": "web_search request failed"}
-        return r.json().get("result", "No result.")
-
-    async def _fetch_remote_content(
-        url: str,
-        max_chars: int = 40000,
-        use_vision_for_images: bool = True,
-        target: str = "url",
-    ) -> Any:
-        raw_url = str(url or "").strip()
-        parsed = urlparse(raw_url)
-        if parsed.scheme not in {"http", "https"}:
-            return {"error": "url must start with http:// or https://"}
-
-        safe_max_chars = max(2000, min(int(max_chars), 120000))
-        try:
-            async with httpx.AsyncClient(
-                timeout=45.0,
-                follow_redirects=True,
-                headers={"User-Agent": "OpenTulpa/0.1 (+content-fetch)"},
-            ) as client:
-                resp = await client.get(raw_url)
-        except Exception as exc:
-            return {"error": f"link fetch failed: {exc}"}
-
-        if resp.status_code >= 400:
-            return {"error": f"link fetch failed: HTTP {resp.status_code}"}
-
-        ctype = str(resp.headers.get("content-type", "")).split(";")[0].strip().lower()
-        final_url = str(resp.url)
-        text_content = ""
-        title: str | None = None
-        mode = "text"
-        is_image = ctype.startswith("image/")
-        is_pdf = ctype == "application/pdf" or final_url.lower().endswith(".pdf")
-        is_docx = (
-            ctype == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
-            or final_url.lower().endswith(".docx")
-        )
-        file_like = is_image or is_pdf or is_docx
-
-        safe_target = str(target or "url").strip().lower()
-        if safe_target == "url" and file_like:
-            return {
-                "error": (
-                    "URL points to a file-like resource (image/pdf/docx). "
-                    "Use fetch_file_content instead."
-                ),
-                "url": final_url,
-                "content_type": ctype or "unknown",
-            }
-        if safe_target == "file" and not file_like:
-            return {
-                "error": (
-                    "URL does not point to supported file-like content (image/pdf/docx). "
-                    "Use fetch_url_content instead."
-                ),
-                "url": final_url,
-                "content_type": ctype or "unknown",
-            }
-
-        try:
-            if is_image:
-                mode = "image_vision"
-                if use_vision_for_images:
-                    vision = await runtime._model.ainvoke(
-                        [
-                            SystemMessage(
-                                content=(
-                                    "Describe the image and extract all readable text. "
-                                    "If it is a screenshot/document, summarize key points."
-                                )
-                            ),
-                            HumanMessage(
-                                content=[
-                                    {"type": "text", "text": "Analyze this image URL."},
-                                    {"type": "image_url", "image_url": {"url": final_url}},
-                                ]
-                            ),
-                        ]
-                    )
-                    text_content = _content_to_text(getattr(vision, "content", "")).strip()
-            elif is_pdf:
-                mode = "pdf_llm"
-                text_content = await summarize_uploaded_blob(
-                    runtime,
-                    filename=final_url.rsplit("/", 1)[-1] or "document.pdf",
-                    mime_type=ctype or "application/pdf",
-                    kind="document",
-                    raw_bytes=resp.content,
-                    question=(
-                        "Extract key information from this PDF and provide a concise but complete "
-                        "summary with important facts, entities, dates, and actions."
-                    ),
-                )
-            elif is_docx:
-                mode = "docx_llm"
-                text_content = await summarize_uploaded_blob(
-                    runtime,
-                    filename=final_url.rsplit("/", 1)[-1] or "document.docx",
-                    mime_type=ctype
-                    or "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-                    kind="document",
-                    raw_bytes=resp.content,
-                    question=(
-                        "Extract key information from this DOCX and provide a concise but complete "
-                        "summary with important facts, entities, dates, and actions."
-                    ),
-                )
-            else:
-                mode = "web_text"
-                raw_text = resp.text
-                if "html" in ctype or "<html" in raw_text.lower():
-                    crawled_text, crawled_title, crawl_error = await _crawl4ai_extract(final_url)
-                    if crawled_text:
-                        mode = "web_text_crawl4ai"
-                        title = crawled_title or _extract_html_title(raw_text)
-                        text_content = crawled_text
-                    else:
-                        mode = "web_text_fallback"
-                        title = _extract_html_title(raw_text)
-                        text_content = _html_to_text(raw_text)
-                        if crawl_error:
-                            text_content = (
-                                f"[crawl4ai_fallback_reason] {crawl_error}\n\n{text_content}"
-                            )
-                else:
-                    text_content = raw_text
-        except Exception as exc:
-            return {
-                "error": f"content extraction failed: {exc}",
-                "url": final_url,
-                "content_type": ctype or "unknown",
-            }
-
-        normalized = re.sub(r"\n{3,}", "\n\n", str(text_content or "").strip())
-        truncated = len(normalized) > safe_max_chars
-        return {
-            "url": final_url,
-            "content_type": ctype or "unknown",
-            "mode": mode,
-            "title": title,
-            "char_count": len(normalized),
-            "truncated": truncated,
-            "text": normalized[:safe_max_chars],
-        }
-
-    @tool
-    async def fetch_url_content(url: str, max_chars: int = 40000) -> Any:
-        """Fetch and extract web page/text/JSON content from a URL."""
-        return await _fetch_remote_content(
-            url=url,
-            max_chars=max_chars,
-            use_vision_for_images=False,
-            target="url",
-        )
-
-    @tool
-    async def fetch_file_content(
-        url: str,
-        max_chars: int = 40000,
-        use_vision_for_images: bool = True,
-    ) -> Any:
-        """Fetch and analyze file-like URL content (image/pdf/docx)."""
-        return await _fetch_remote_content(
-            url=url,
-            max_chars=max_chars,
-            use_vision_for_images=use_vision_for_images,
-            target="file",
-        )
-
-    @tool
-    async def tulpa_write_file(path: str, content: str) -> Any:
-        """Write file in approved paths."""
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/tulpa/write_file",
-            json_body={"path": path, "content": content},
-            timeout=20.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"write failed: {r.text}"}
-        return r.json()
-
-    @tool
-    async def tulpa_validate_file(path: str) -> Any:
-        """Validate generated file syntax/contracts in approved paths."""
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/tulpa/validate_file",
-            json_body={"path": path},
-            timeout=20.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"validation failed: {r.text}"}
-        return r.json()
-
-    @tool
-    async def tulpa_reload() -> Any:
-        """Reload tulpa_stuff routers so newly written connectors become active."""
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/tulpa/reload",
-            timeout=20.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"reload failed: {r.text}"}
-        return r.json()
-
-    @tool
-    async def tulpa_run_terminal(
-        command: str,
-        working_dir: str = "tulpa_stuff",
-        timeout_seconds: int = 90,
-        thread_id: str = "",
-        execution_origin: str | None = None,
-    ) -> Any:
-        """Run executable shell/script command in the agent venv."""
-        safe_working_dir = str(working_dir or "").strip() or "tulpa_stuff"
-        safe_command = normalize_command_for_working_dir(
-            command=str(command or "").strip(),
-            working_dir=safe_working_dir,
-        )
-        if not _looks_like_shell_command(safe_command):
-            return {
-                "error": (
-                    "Command rejected: provide a concrete shell command (executable + args), "
-                    "not natural language."
-                )
-            }
-        safe_timeout = max(5, min(int(timeout_seconds), 600))
-        require_customer_id(runtime)
-        safe_thread = str(thread_id or "").strip()
-        normalized_origin = normalize_execution_origin(
-            thread_id=safe_thread,
-            execution_origin=execution_origin,
-        )
-
-        r = await runtime._request_with_backoff(
-            "POST",
-            "/internal/tulpa/run_terminal",
-            json_body={
-                "command": safe_command,
-                "working_dir": safe_working_dir,
-                "timeout_seconds": safe_timeout,
-            },
-            timeout=max(10.0, float(safe_timeout) + 10.0),
-            retries=1,
-        )
-        if r.status_code != 200:
-            return {"error": f"terminal failed: {r.text}"}
-        payload = r.json()
-        if isinstance(payload, dict):
-            payload["execution_origin"] = normalized_origin
-            payload = _decorate_python_dependency_failure(payload)
-        return payload
-
-    @tool
-    async def tulpa_read_file(path: str, max_chars: int = 12000) -> Any:
-        """Read file in approved paths."""
-        safe_max_chars = max(500, min(int(max_chars), 20000))
-        r = await runtime._request_with_backoff(
-            "GET",
-            "/internal/tulpa/read_file",
-            params={"path": path, "max_chars": safe_max_chars},
-            timeout=15.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"read failed: {r.text}"}
-        return r.json()
-
-    @tool
-    async def tulpa_catalog() -> Any:
-        """Get catalog of tracked files and artifacts."""
-        r = await runtime._request_with_backoff("GET", "/internal/tulpa/catalog", timeout=10.0)
-        if r.status_code != 200:
-            return {"error": f"catalog failed: {r.text}"}
-        return r.json().get("catalog", {})
-
-    @tool
-    async def task_status(task_id: str) -> Any:
-        """Get task status."""
-        r = await runtime._request_with_backoff("GET", f"/internal/tasks/{task_id}", timeout=10.0)
-        if r.status_code != 200:
-            return {"error": f"task_status failed: {r.text}"}
-        return r.json().get("task", {})
-
-    @tool
-    async def task_events(task_id: str, limit: int = 30, offset: int = 0) -> Any:
-        """Get task events."""
-        r = await runtime._request_with_backoff(
-            "GET",
-            f"/internal/tasks/{task_id}/events",
-            params={"limit": max(1, min(int(limit), 200)), "offset": max(0, int(offset))},
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"task_events failed: {r.text}"}
-        return r.json().get("events", [])
-
-    @tool
-    async def task_artifacts(task_id: str) -> Any:
-        """Get task artifacts."""
-        r = await runtime._request_with_backoff(
-            "GET", f"/internal/tasks/{task_id}/artifacts", timeout=10.0
-        )
-        if r.status_code != 200:
-            return {"error": f"task_artifacts failed: {r.text}"}
-        return r.json().get("artifacts", [])
-
-    @tool
-    async def task_relaunch(
-        task_id: str, clarification: str | None = None, trigger_reason: str = "user_requested"
-    ) -> Any:
-        """Relaunch a task."""
-        r = await runtime._request_with_backoff(
-            "POST",
-            f"/internal/tasks/{task_id}/relaunch",
-            json_body={"clarification": clarification, "trigger_reason": trigger_reason},
-            timeout=20.0,
-        )
-        if r.status_code != 200:
-            return {"error": f"task_relaunch failed: {r.text}"}
-        return r.json().get("task", {})
-
-    @tool
-    async def task_cancel(task_id: str) -> Any:
-        """Cancel a task."""
-        r = await runtime._request_with_backoff(
-            "POST", f"/internal/tasks/{task_id}/cancel", timeout=10.0
-        )
-        if r.status_code != 200:
-            return {"error": f"task_cancel failed: {r.text}"}
-        return r.json().get("task", {})
-
-    @tool
-    async def server_time() -> Any:
-        """Get server time."""
-        now_local = datetime.now().astimezone()
-        now_utc = datetime.now(UTC)
-        return {
-            "server_time_local_iso": now_local.isoformat(),
-            "server_timezone": str(now_local.tzinfo),
-            "server_time_utc_iso": now_utc.isoformat(),
-            "unix_timestamp": int(now_utc.timestamp()),
-        }
-
-    return {
-        "send_owner_update": send_owner_update,
-        "memory_search": memory_search,
-        "memory_add": memory_add,
-        "uploaded_file_search": uploaded_file_search,
-        "uploaded_file_get": uploaded_file_get,
-        "uploaded_file_send": uploaded_file_send,
-        "tulpa_file_send": tulpa_file_send,
-        "web_image_send": web_image_send,
-        "uploaded_file_analyze": uploaded_file_analyze,
-        "uploaded_file_inspect_structure": uploaded_file_inspect_structure,
-        "business_knowledge_index": business_knowledge_index,
-        "business_knowledge_query": business_knowledge_query,
-        "user_context_add_files": user_context_add_files,
-        "user_context_query": user_context_query,
-        "user_context_list_sources": user_context_list_sources,
-        "user_context_find_sources": user_context_find_sources,
-        "user_context_reindex": user_context_reindex,
-        "user_context_archive_sources": user_context_archive_sources,
-        "user_context_promote_to_intake": user_context_promote_to_intake,
-        "directive_get": directive_get,
-        "directive_set": directive_set,
-        "directive_clear": directive_clear,
-        "time_profile_get": time_profile_get,
-        "time_profile_set": time_profile_set,
-        "web_search": web_search,
-        "fetch_url_content": fetch_url_content,
-        "fetch_file_content": fetch_file_content,
-        "tulpa_write_file": tulpa_write_file,
-        "tulpa_validate_file": tulpa_validate_file,
-        "tulpa_reload": tulpa_reload,
-        "tulpa_run_terminal": tulpa_run_terminal,
-        "tulpa_read_file": tulpa_read_file,
-        "tulpa_catalog": tulpa_catalog,
-        "task_status": task_status,
-        "task_events": task_events,
-        "task_artifacts": task_artifacts,
-        "task_relaunch": task_relaunch,
-        "task_cancel": task_cancel,
-        "server_time": server_time,
-    }
+    tools: dict[str, Any] = {}
+    tools.update(register_owner_update_tools(runtime))
+    tools.update(register_memory_tools(runtime))
+    tools.update(register_file_tools(runtime))
+    tools.update(register_business_knowledge_tools(runtime))
+    tools.update(register_user_context_tools(runtime))
+    tools.update(register_directive_tools(runtime))
+    tools.update(register_time_profile_tools(runtime))
+    tools.update(register_web_tools(runtime))
+    tools.update(register_tulpa_workspace_tools(runtime))
+    tools.update(register_task_tools(runtime))
+    tools.update(register_server_time_tools(runtime))
+    return tools
