+++ src/opentulpa/agent/tools/task_tools.py
@@ -1,0 +1,73 @@
+"""Task tools."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from langchain.tools import tool
+
+
+def register_task_tools(runtime: Any) -> dict[str, Any]:
+    @tool
+    async def task_status(task_id: str) -> Any:
+        """Get task status."""
+        r = await runtime._request_with_backoff("GET", f"/internal/tasks/{task_id}", timeout=10.0)
+        if r.status_code != 200:
+            return {"error": f"task_status failed: {r.text}"}
+        return r.json().get("task", {})
+
+    @tool
+    async def task_events(task_id: str, limit: int = 30, offset: int = 0) -> Any:
+        """Get task events."""
+        r = await runtime._request_with_backoff(
+            "GET",
+            f"/internal/tasks/{task_id}/events",
+            params={"limit": max(1, min(int(limit), 200)), "offset": max(0, int(offset))},
+            timeout=10.0,
+        )
+        if r.status_code != 200:
+            return {"error": f"task_events failed: {r.text}"}
+        return r.json().get("events", [])
+
+    @tool
+    async def task_artifacts(task_id: str) -> Any:
+        """Get task artifacts."""
+        r = await runtime._request_with_backoff(
+            "GET", f"/internal/tasks/{task_id}/artifacts", timeout=10.0
+        )
+        if r.status_code != 200:
+            return {"error": f"task_artifacts failed: {r.text}"}
+        return r.json().get("artifacts", [])
+
+    @tool
+    async def task_relaunch(
+        task_id: str, clarification: str | None = None, trigger_reason: str = "user_requested"
+    ) -> Any:
+        """Relaunch a task."""
+        r = await runtime._request_with_backoff(
+            "POST",
+            f"/internal/tasks/{task_id}/relaunch",
+            json_body={"clarification": clarification, "trigger_reason": trigger_reason},
+            timeout=20.0,
+        )
+        if r.status_code != 200:
+            return {"error": f"task_relaunch failed: {r.text}"}
+        return r.json().get("task", {})
+
+    @tool
+    async def task_cancel(task_id: str) -> Any:
+        """Cancel a task."""
+        r = await runtime._request_with_backoff(
+            "POST", f"/internal/tasks/{task_id}/cancel", timeout=10.0
+        )
+        if r.status_code != 200:
+            return {"error": f"task_cancel failed: {r.text}"}
+        return r.json().get("task", {})
+
+    return {
+        "task_status": task_status,
+        "task_events": task_events,
+        "task_artifacts": task_artifacts,
+        "task_relaunch": task_relaunch,
+        "task_cancel": task_cancel,
+    }
